Create files for classes
